requirements.txt

Django==3.1.8

python-dateutil==2.8.1
redis==3.5.3
martor
gunicorn
